feat: reorganize file's structure
