+++ my-app/src/presentational_components/movieCard/movieCard.scss
