+++ my-app/src/presentational_components/searchResultString/searchResult.jsx
