+++ my-app/src/shared_components/footer/footer.scss
