+++ my-app/src/shared_components/panicBtn/panicButton.scss
