+++ my-app/src/shared_components/searchBy/search.scss
